Добавлена реализация проверки апи между бэком и дб

+++ src/main/java/redslicedatabase/redslicedatabase/controller/BranchController.java
@@ -9,8 +9,10 @@ import jakarta.validation.Valid;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import redslicedatabase.redslicedatabase.config.AppConfig;
 import redslicedatabase.redslicedatabase.dto.BranchDTO.inbound.CreateBranchDTO;
 import redslicedatabase.redslicedatabase.dto.BranchDTO.outbound.BranchDTO;
 import redslicedatabase.redslicedatabase.model.Branch;
@@ -19,6 +21,8 @@ import redslicedatabase.redslicedatabase.service.ChatService;
 
 import java.nio.file.AccessDeniedException;
 import java.util.List;
+import java.util.Map;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/branches")
@@ -30,10 +34,20 @@ public class BranchController {
     private BranchService branchService;
     @Autowired
     private ChatService chatService;
+    @Autowired
+    private AppConfig appConfig;
 
     // Создать новую ветку
     @PostMapping
-    public ResponseEntity<BranchDTO> createBranch(@Valid @RequestBody CreateBranchDTO createBranchDTO) {
+    public ResponseEntity<?> createBranch(@Valid @RequestBody CreateBranchDTO createBranchDTO,
+                                                  @RequestHeader String apiDBKey) {
+        if (!Objects.equals(apiDBKey, appConfig.getApiDatabaseKey())) {
+            return ResponseEntity
+                    .status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body(Map.of(
+                            "error", "Wrong apiDBKey"
+                    ));
+        }
 
         Branch createdBranch = branchService.createBranch(createBranchDTO);
         return ResponseEntity.ok(branchService.convertToDTO(createdBranch));
@@ -41,8 +55,17 @@ public class BranchController {
 
     // Получение всех веток чата с валидацией пользователя
     @GetMapping("/chat/{chatId}/validate")
-    public ResponseEntity<List<BranchDTO>> getBranchByChatIdAndFirebase(@PathVariable Long chatId,
-                                                                        @RequestParam String uidFirebase) throws AccessDeniedException {
+    public ResponseEntity<?> getBranchByChatIdAndFirebase(@PathVariable Long chatId,
+                                                                        @RequestParam String uidFirebase,
+                                                                        @RequestHeader String apiDBKey) throws AccessDeniedException {
+        if (!Objects.equals(apiDBKey, appConfig.getApiDatabaseKey())) {
+            return ResponseEntity
+                    .status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body(Map.of(
+                            "error", "Wrong apiDBKey"
+                    ));
+        }
+
         chatService.getChatByIdWithAccessCheck(chatId, uidFirebase); // Проверка на доступность чата пользователю
 
         logger.info("GET UID: Got chat id: {}", chatId);
@@ -55,8 +78,17 @@ public class BranchController {
 
     // Каскадное удаление ветки с проверкой доступа
     @DeleteMapping("/{id}/validate")
-    public ResponseEntity<Void> deleteBranchById(@PathVariable Long id,
-                                                 @RequestParam String uidFirebase) throws AccessDeniedException {
+    public ResponseEntity<?> deleteBranchById(@PathVariable Long id,
+                                                 @RequestParam String uidFirebase,
+                                                 @RequestHeader String apiDBKey) throws AccessDeniedException {
+        if (!Objects.equals(apiDBKey, appConfig.getApiDatabaseKey())) {
+            return ResponseEntity
+                    .status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body(Map.of(
+                            "error", "Wrong apiDBKey"
+                    ));
+        }
+
         branchService.deleteBranchByIdWithAccessCheck(id, uidFirebase);
         return ResponseEntity.noContent().build(); // Возвращает 204 No Content
     }
